Refactor the provider
