+++ src/App.js
@@ -1,6 +1,5 @@
 import { useState, useEffect } from 'react';
 import { Route, Routes } from 'react-router-dom';
-import UserContext from './context/UserContext';
 
 // App Components
 import Header from './components/Header';
@@ -11,8 +10,6 @@ import Settings from './components/Settings';
 import NotFound from './components/NotFound';
 
 function App() {
-  const [user, setUser] = useState(null);
-
   const [isDarkMode, setIsDarkMode] = useState(false);
   const [accentColor, setAccentColor] = useState('#63537d');
   const [fontPercentage, setFontPercentage] = useState(100);
@@ -26,59 +23,37 @@ function App() {
     document.body.style.fontSize = `${fontPercentage}%`;
   }, [isDarkMode, fontPercentage]);
 
-  const signInUser = (username, password) => {
-    const newUser = {
-      username,
-      password,
-    };
-    setUser(newUser);
-  };
-
-  const signOutUser = () => {
-    setUser(null);
-  };
-
   const toggleDarkMode = () => {
     setIsDarkMode((currentMode) => !currentMode);
   };
 
   return (
-    <UserContext.Provider
-      value={{
-        user,
-        actions: {
-          signIn: signInUser,
-          signOut: signOutUser,
-        },
-      }}
-    >
-      <div>
-        {/* {console.log(UserContext)} */}
-        <Header accentColor={accentColor} />
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route
-            path="signin"
-            element={<UserSignIn accentColor={accentColor} />}
-          />
-          <Route path="signout" element={<UserSignOut />} />
-          <Route
-            path="settings"
-            element={
-              <Settings
-                isDarkMode={isDarkMode}
-                toggleDarkMode={toggleDarkMode}
-                accentColor={accentColor}
-                updateAccentColor={setAccentColor}
-                fontPercentage={fontPercentage}
-                updateFontPercentage={setFontPercentage}
-              />
-            }
-          />
-          <Route path="*" element={<NotFound />} />
-        </Routes>
-      </div>
-    </UserContext.Provider>
+    <div>
+      {/* {console.log(UserContext)} */}
+      <Header accentColor={accentColor} />
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route
+          path="signin"
+          element={<UserSignIn accentColor={accentColor} />}
+        />
+        <Route path="signout" element={<UserSignOut />} />
+        <Route
+          path="settings"
+          element={
+            <Settings
+              isDarkMode={isDarkMode}
+              toggleDarkMode={toggleDarkMode}
+              accentColor={accentColor}
+              updateAccentColor={setAccentColor}
+              fontPercentage={fontPercentage}
+              updateFontPercentage={setFontPercentage}
+            />
+          }
+        />
+        <Route path="*" element={<NotFound />} />
+      </Routes>
+    </div>
   );
 }
